08_REST_ManyToMany_CRUD/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.5.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.12
django-bootstrap-v5==1.0.5
django-extensions==3.1.3
django-seed==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.3
executing==0.8.3
Faker==13.3.4
idna==3.3
inflection==0.5.1
ipykernel==6.7.0
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
packaging==21.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.20
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.12.1
testpath==0.5.0
tomli==1.2.3
toposort==1.7
tornado==6.1
traitlets==5.0.5
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
